Add support for built applications in build script
A new sibling directory to apps/ - built-apps/ - is introduced to contain already built applications to be copied to apps/ during the build process.
All applications in built-apps/ are neither npm installed or mopped because they are ready to be deployed.

This provide us enough flexibility to have applications with slight changes to their original source code or dependencies.